requirements.txt
anthropic>=0.64.0
httpx>=0.27.0
cryptography>=43.0.0
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.0.0
